[NTUSER] Preserve pointer-sized system timer IDs
TIMERPROC receives its timer identifier as UINT_PTR, but the WM_SYSTIMER dispatch path narrowed MSG.wParam to UINT before invoking the callback. Pass the pointer-sized value through unchanged on 64-bit systems.

+++ win32ss/user/ntuser/message.c
@@ -934,7 +934,7 @@ IntDispatchMessage(PMSG pMsg)
             if (pTimer && pTimer->pfn)
             {
                 Time = EngGetTickCount32();
-                pTimer->pfn(pMsg->hwnd, WM_SYSTIMER, (UINT)pMsg->wParam, Time);
+                pTimer->pfn(pMsg->hwnd, WM_SYSTIMER, pMsg->wParam, Time);
             }
             return 0;
         }
